+++ packages/zowe-explorer/__tests__/__unit__/trees/uss/UssFSProvider.unit.test.ts
@@ -10,7 +10,18 @@
  */
 
 import { Disposable, FilePermission, FileSystemError, FileType, TextEditor, Uri, workspace } from "vscode";
-import { AuthHandler, BaseProvider, DirEntry, FileEntry, Gui, UssDirectory, UssFile, ZoweExplorerApiType, ZoweScheme } from "@zowe/zowe-explorer-api";
+import {
+    AuthHandler,
+    BaseProvider,
+    DirEntry,
+    FileEntry,
+    FsAbstractUtils,
+    Gui,
+    UssDirectory,
+    UssFile,
+    ZoweExplorerApiType,
+    ZoweScheme,
+} from "@zowe/zowe-explorer-api";
 import { Profiles } from "../../../../src/configuration/Profiles";
 import { createIProfile } from "../../../__mocks__/mockCreators/shared";
 import { ZoweExplorerApiRegister } from "../../../../src/extending/ZoweExplorerApiRegister";
@@ -20,6 +31,7 @@ import { MockedProperty } from "../../../__mocks__/mockUtils";
 import { ZoweLogger } from "../../../../src/tools/ZoweLogger";
 import { ProfilesUtils } from "../../../../src/utils/ProfilesUtils";
 import { AuthUtils } from "../../../../src/utils/AuthUtils";
+import * as vscode from "vscode";
 
 const testProfile = createIProfile();
 
@@ -138,6 +150,15 @@ describe("UssFSProvider", () => {
         it("returns a file as-is when query has inDiff parameter", async () => {
             lookupMock.mockReturnValueOnce(testEntries.file);
             await expect(UssFSProvider.instance.stat(testUris.file.with({ query: "inDiff=true" }))).resolves.toStrictEqual(testEntries.file);
+            expect(lookupMock).toHaveBeenCalledWith(testUris.file, false);
+        });
+
+        it("looks up the resource before loading profile which may fail", async () => {
+            const lookupMock = jest.spyOn((UssFSProvider as any).prototype, "lookup").mockReturnValueOnce(testEntries.file);
+            jest.spyOn(FsAbstractUtils, "getInfoForUri").mockImplementationOnce(() => {
+                throw new Error("invalid profile");
+            });
+            await expect(UssFSProvider.instance.stat(testUris.file)).rejects.toThrow("invalid profile");
             expect(lookupMock).toHaveBeenCalledWith(testUris.file, false);
         });
     });
@@ -273,7 +294,7 @@ describe("UssFSProvider", () => {
                     commandResponse: "",
                 });
                 const lookupParentDirMock = jest
-                    .spyOn(UssFSProvider.instance as any, "_lookupParentDirectory")
+                    .spyOn(UssFSProvider.instance as any, "lookupParentDirectory")
                     .mockReturnValueOnce(null)
                     .mockReturnValueOnce({ ...testEntries.folder, entries: new Map() });
                 const createDirMock = jest.spyOn(workspace.fs, "createDirectory").mockImplementation();
@@ -346,16 +367,21 @@ describe("UssFSProvider", () => {
     });
 
     describe("fetchFileAtUri", () => {
-        beforeEach(() => {
-            Object.defineProperty(Profiles, "getInstance", {
-                value: jest.fn(() => {
-                    return {
-                        loadNamedProfile: jest.fn(() => {
-                            return testProfile;
-                        }),
-                    };
-                }),
-            });
+        let mockedProfilesProp: MockedProperty;
+        beforeAll(() => {
+            mockedProfilesProp = new MockedProperty(
+                Profiles,
+                "getInstance",
+                undefined,
+                jest.fn().mockReturnValue({
+                    loadNamedProfile: jest.fn(() => {
+                        return testProfile;
+                    }),
+                })
+            );
+        });
+        afterAll(() => {
+            mockedProfilesProp[Symbol.dispose]();
         });
         it("calls getContents to get the data for a file entry", async () => {
             const fileEntry = { ...testEntries.file };
@@ -626,7 +652,7 @@ describe("UssFSProvider", () => {
             const lookupAsFileMock = jest.spyOn(UssFSProvider.instance as any, "_lookupAsFile").mockImplementationOnce((uri) => {
                 throw FileSystemError.FileNotFound(uri as Uri);
             });
-            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(testEntries.folder);
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(testEntries.folder);
 
             let err;
             try {
@@ -644,7 +670,7 @@ describe("UssFSProvider", () => {
             const lookupAsFileMock = jest.spyOn(UssFSProvider.instance as any, "_lookupAsFile").mockImplementationOnce((uri) => {
                 throw FileSystemError.FileNotFound(uri as Uri);
             });
-            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(null);
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(null);
             const remoteLookupForResource = jest
                 .spyOn(UssFSProvider.instance, "remoteLookupForResource")
                 .mockResolvedValueOnce(testEntries.folder as any);
@@ -722,7 +748,7 @@ describe("UssFSProvider", () => {
                 }),
             };
 
-            ProfilesUtils.extenderTypeReady.set(testProfile.name, profilePromise);
+            ProfilesUtils.extenderTypeReady.set(testProfile.name || "mockProfile", profilePromise as any);
 
             const lookupAsFileMock = jest.spyOn(UssFSProvider.instance as any, "_lookupAsFile");
             lookupAsFileMock.mockReturnValue(testEntries.file);
@@ -765,12 +791,16 @@ describe("UssFSProvider", () => {
 
             const resolveProfile = jest.fn();
             const profilePromise = {
+                mStatus: "pending",
+                resolve: jest.fn(),
+                reject: jest.fn(),
+                status: "pending",
                 promise: new Promise<void>((resolve) => {
                     resolveProfile.mockImplementation(resolve);
                     setTimeout(resolve, 50);
                 }),
             };
-            jest.spyOn(ProfilesUtils.extenderTypeReady, "get").mockReturnValueOnce(profilePromise);
+            jest.spyOn(ProfilesUtils.extenderTypeReady, "get").mockReturnValueOnce(profilePromise as any);
             const lookupAsFileMock = jest.spyOn(UssFSProvider.instance as any, "_lookupAsFile");
             lookupAsFileMock.mockReturnValue(testEntries.file);
 
@@ -810,7 +840,7 @@ describe("UssFSProvider", () => {
                 ...testEntries.folder,
                 entries: new Map([[testEntries.file.name, { ...testEntries.file }]]),
             };
-            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(folder);
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(folder);
             const newContents = new Uint8Array([3, 6, 9]);
             await UssFSProvider.instance.writeFile(testUris.file, newContents, { create: false, overwrite: true });
 
@@ -838,7 +868,7 @@ describe("UssFSProvider", () => {
                 ...testEntries.folder,
                 entries: new Map([[testEntries.file.name, { ...testEntries.file }]]),
             };
-            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(folder);
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(folder);
             const autoDetectEncodingMock = jest.spyOn(UssFSProvider.instance, "autoDetectEncoding").mockResolvedValue(undefined);
             const newContents = new Uint8Array([3, 6, 9]);
             const handleConflictMock = jest.spyOn(UssFSProvider.instance as any, "_handleConflict").mockImplementation();
@@ -871,7 +901,7 @@ describe("UssFSProvider", () => {
                 ...testEntries.folder,
                 entries: new Map([[testEntries.file.name, { ...testEntries.file }]]),
             };
-            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(folder);
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(folder);
             const autoDetectEncodingMock = jest.spyOn(UssFSProvider.instance, "autoDetectEncoding").mockResolvedValue(undefined);
             const newContents = new Uint8Array([3, 6, 9]);
             const handleConflictMock = jest.spyOn(UssFSProvider.instance as any, "_handleConflict").mockImplementation();
@@ -905,7 +935,7 @@ describe("UssFSProvider", () => {
                 ...testEntries.session,
                 entries: new Map(),
             };
-            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(folder);
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(folder);
             const autoDetectEncodingMock = jest.spyOn(UssFSProvider.instance, "autoDetectEncoding").mockResolvedValue(undefined);
             const newContents = new Uint8Array([3, 6, 9]);
             await UssFSProvider.instance.writeFile(testUris.file, newContents, { create: true, overwrite: true });
@@ -931,7 +961,7 @@ describe("UssFSProvider", () => {
                 ...testEntries.folder,
                 entries: new Map([[testEntries.file.name, { ...testEntries.file, wasAccessed: false }]]),
             };
-            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(folder);
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(folder);
             const newContents = new Uint8Array([]);
             await UssFSProvider.instance.writeFile(testUris.file, newContents, { create: false, overwrite: true });
 
@@ -947,7 +977,7 @@ describe("UssFSProvider", () => {
                 ...testEntries.folder,
                 entries: new Map([[testEntries.file.name, { ...testEntries.file, wasAccessed: false }]]),
             };
-            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(folder);
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(folder);
             const newContents = new Uint8Array([]);
             await UssFSProvider.instance.writeFile(
                 testUris.file.with({
@@ -980,7 +1010,7 @@ describe("UssFSProvider", () => {
                 ...testEntries.session,
                 entries: new Map([[testEntries.file.name, { ...testEntries.file, wasAccessed: false }]]),
             };
-            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(rootFolder);
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(rootFolder);
             let err;
             try {
                 await UssFSProvider.instance.writeFile(testUris.file, new Uint8Array([]), { create: true, overwrite: false });
@@ -997,7 +1027,7 @@ describe("UssFSProvider", () => {
                 ...testEntries.session,
                 entries: new Map([[testEntries.folder.name, { ...testEntries.folder }]]),
             };
-            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(rootFolder);
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(rootFolder);
             let err;
             try {
                 await UssFSProvider.instance.writeFile(testUris.folder, new Uint8Array([]), { create: true, overwrite: false });
@@ -1013,7 +1043,7 @@ describe("UssFSProvider", () => {
     describe("makeEmptyFileWithEncoding", () => {
         it("creates an empty file in the provider with the given encoding", () => {
             const fakeSession = { ...testEntries.session };
-            const parentDirMock = jest.spyOn(UssFSProvider.instance as any, "_lookupParentDirectory").mockReturnValueOnce(fakeSession);
+            const parentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValueOnce(fakeSession);
             expect(UssFSProvider.instance.makeEmptyFileWithEncoding(testUris.file, { kind: "binary" }));
             expect(fakeSession.entries.has(testEntries.file.name)).toBe(true);
             parentDirMock.mockRestore();
@@ -1021,27 +1051,21 @@ describe("UssFSProvider", () => {
     });
 
     describe("rename", () => {
-        it("throws an error if entry exists and 'overwrite' is false", async () => {
-            const lookupMock = jest.spyOn(UssFSProvider.instance as any, "lookup").mockReturnValueOnce({ ...testEntries.file });
-            await expect(
-                UssFSProvider.instance.rename(testUris.file, testUris.file.with({ path: "/sestest/aFile2.txt" }), { overwrite: false })
-            ).rejects.toThrow("Rename failed: aFile2.txt already exists in /");
-
-            lookupMock.mockRestore();
-        });
-
         it("renames a file entry in the FSP and remote system", async () => {
             const mockUssApi = {
                 rename: jest.fn(),
             };
             const ussApiMock = jest.spyOn(ZoweExplorerApiRegister, "getUssApi").mockReturnValueOnce(mockUssApi as any);
-            const lookupMock = jest.spyOn(UssFSProvider.instance as any, "lookup").mockReturnValueOnce({ ...testEntries.file });
             const fileEntry = { ...testEntries.file, metadata: { ...testEntries.file.metadata } };
             const sessionEntry = {
                 ...testEntries.session,
                 entries: new Map([[testEntries.file.name, fileEntry]]),
             };
             (UssFSProvider.instance as any).root.entries.set("sestest", sessionEntry);
+            const lookupMock = jest.spyOn(UssFSProvider.instance as any, "lookup").mockImplementation((uri: any) => {
+                return sessionEntry.entries.get("aFile.txt");
+            });
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValue(sessionEntry);
 
             await UssFSProvider.instance.rename(testUris.file, testUris.file.with({ path: "/sestest/aFile2.txt" }), { overwrite: true });
             expect(mockUssApi.rename).toHaveBeenCalledWith("/aFile.txt", "/aFile2.txt");
@@ -1049,6 +1073,7 @@ describe("UssFSProvider", () => {
             expect(sessionEntry.entries.has("aFile2.txt")).toBe(true);
 
             lookupMock.mockRestore();
+            lookupParentDirMock.mockRestore();
             ussApiMock.mockRestore();
         });
 
@@ -1057,14 +1082,17 @@ describe("UssFSProvider", () => {
                 rename: jest.fn(),
             };
             const ussApiMock = jest.spyOn(ZoweExplorerApiRegister, "getUssApi").mockReturnValueOnce(mockUssApi as any);
-            const lookupMock = jest.spyOn(UssFSProvider.instance as any, "lookup").mockReturnValueOnce({ ...testEntries.folder });
-            const updChildPathsMock = jest.spyOn(UssFSProvider.instance as any, "_updateChildPaths").mockResolvedValueOnce(undefined);
             const folderEntry = { ...testEntries.folder, metadata: { ...testEntries.folder.metadata } };
             const sessionEntry = {
                 ...testEntries.session,
                 entries: new Map([[testEntries.folder.name, folderEntry]]),
             };
             (UssFSProvider.instance as any).root.entries.set("sestest", sessionEntry);
+            const lookupMock = jest.spyOn(UssFSProvider.instance as any, "lookup").mockImplementation((uri: any) => {
+                return sessionEntry.entries.get("aFolder");
+            });
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValue(sessionEntry);
+            const updChildPathsMock = jest.spyOn(UssFSProvider.instance as any, "_updateChildPaths").mockResolvedValueOnce(undefined);
 
             await UssFSProvider.instance.rename(testUris.folder, testUris.folder.with({ path: "/sestest/aFolder2" }), { overwrite: true });
             expect(mockUssApi.rename).toHaveBeenCalledWith("/aFolder", "/aFolder2");
@@ -1073,6 +1101,7 @@ describe("UssFSProvider", () => {
             expect(updChildPathsMock).toHaveBeenCalledWith(folderEntry);
 
             lookupMock.mockRestore();
+            lookupParentDirMock.mockRestore();
             ussApiMock.mockRestore();
             updChildPathsMock.mockRestore();
         });
@@ -1083,13 +1112,16 @@ describe("UssFSProvider", () => {
             };
             const errMsgSpy = jest.spyOn(Gui, "errorMessage");
             const ussApiMock = jest.spyOn(ZoweExplorerApiRegister, "getUssApi").mockReturnValueOnce(mockUssApi as any);
-            const lookupMock = jest.spyOn(UssFSProvider.instance as any, "lookup").mockReturnValueOnce({ ...testEntries.folder });
             const folderEntry = { ...testEntries.folder, metadata: { ...testEntries.folder.metadata } };
             const sessionEntry = {
                 ...testEntries.session,
                 entries: new Map([[testEntries.folder.name, folderEntry]]),
             };
             (UssFSProvider.instance as any).root.entries.set("sestest", sessionEntry);
+            const lookupMock = jest.spyOn(UssFSProvider.instance as any, "lookup").mockImplementation((uri: any) => {
+                return sessionEntry.entries.get("aFolder");
+            });
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValue(sessionEntry);
 
             await expect(
                 UssFSProvider.instance.rename(testUris.folder, testUris.folder.with({ path: "/sestest/aFolder2" }), { overwrite: true })
@@ -1102,6 +1134,92 @@ describe("UssFSProvider", () => {
             });
 
             lookupMock.mockRestore();
+            lookupParentDirMock.mockRestore();
+            ussApiMock.mockRestore();
+        });
+
+        it("retries rename if FileExists error and listFiles throws 404", async () => {
+            const fileExistsError = Object.assign(vscode.FileSystemError.FileExists("file exists"), { code: "FileExists" });
+
+            const mockUssApiRename = jest.fn().mockRejectedValueOnce(fileExistsError).mockResolvedValueOnce({ success: true });
+
+            const mockUssApi = { rename: mockUssApiRename };
+            jest.spyOn(ZoweExplorerApiRegister, "getUssApi").mockReturnValue(mockUssApi as any);
+
+            const folderEntry = {
+                ...testEntries.folder,
+                entries: new Map(),
+                metadata: { ...testEntries.folder.metadata },
+            };
+
+            const sessionEntry = {
+                ...testEntries.session,
+                entries: new Map([[testEntries.folder.name, folderEntry]]),
+            };
+            (UssFSProvider.instance as any).root.entries.set("sestest", sessionEntry);
+
+            const lookupMock = jest.spyOn(UssFSProvider.instance as any, "lookup").mockImplementation((uri: any) => {
+                return sessionEntry.entries.get("aFolder");
+            });
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValue(sessionEntry);
+
+            const listFilesMock = jest.spyOn(UssFSProvider.instance as any, "listFiles").mockRejectedValue({
+                name: "Error",
+                errorCode: 404,
+            });
+
+            const oldUri = testUris.folder;
+            const newUri = testUris.folder.with({ path: "/sestest/aFolder2" });
+
+            await UssFSProvider.instance.rename(oldUri, newUri, { overwrite: true });
+
+            expect(mockUssApiRename).toHaveBeenCalledTimes(2);
+            expect(mockUssApiRename).toHaveBeenCalledWith("/aFolder", "/aFolder2");
+            expect(sessionEntry.entries.has("aFolder")).toBe(false);
+            expect(sessionEntry.entries.has("aFolder2")).toBe(true);
+
+            listFilesMock.mockRestore();
+            lookupMock.mockRestore();
+            lookupParentDirMock.mockRestore();
+        });
+        it("displays error message if FileExists error occurs and listFiles fails", async () => {
+            const fileExistsError = Object.assign(vscode.FileSystemError.FileExists("File already exists"), { code: "FileExists" });
+
+            const mockUssApi = {
+                rename: jest.fn().mockRejectedValueOnce(fileExistsError),
+            };
+
+            const errMsgSpy = jest.spyOn(Gui, "errorMessage");
+            const ussApiMock = jest.spyOn(ZoweExplorerApiRegister, "getUssApi").mockReturnValue(mockUssApi as any);
+            const listFilesMock = jest.spyOn(UssFSProvider.instance as any, "listFiles").mockResolvedValueOnce({ success: false });
+
+            const folderEntry = {
+                ...testEntries.folder,
+                entries: new Map(),
+                metadata: { ...testEntries.folder.metadata },
+            };
+
+            const sessionEntry = {
+                ...testEntries.session,
+                entries: new Map([[testEntries.folder.name, folderEntry]]),
+            };
+
+            (UssFSProvider.instance as any).root.entries.set("sestest", sessionEntry);
+
+            const lookupMock = jest.spyOn(UssFSProvider.instance as any, "lookup").mockImplementation((uri: any) => {
+                return sessionEntry.entries.get("aFolder");
+            });
+
+            const lookupParentDirMock = jest.spyOn(UssFSProvider.instance as any, "lookupParentDirectory").mockReturnValue(sessionEntry);
+
+            await UssFSProvider.instance.rename(testUris.folder, testUris.folder.with({ path: "/sestest/aFolder2" }), { overwrite: true });
+
+            expect(mockUssApi.rename).toHaveBeenCalledWith("/aFolder", "/aFolder2");
+            expect(errMsgSpy).toHaveBeenCalledWith("File already exists");
+
+            lookupMock.mockRestore();
+            lookupParentDirMock.mockRestore();
+            listFilesMock.mockRestore();
             ussApiMock.mockRestore();
         });
     });
@@ -1661,12 +1779,16 @@ describe("UssFSProvider", () => {
             it("makes API calls when profile is not locked", async () => {
                 const resolveProfile = jest.fn();
                 const profilePromise = {
+                    mStatus: "pending",
+                    resolve: jest.fn(),
+                    reject: jest.fn(),
+                    status: "pending",
                     promise: new Promise<void>((resolve) => {
                         resolveProfile.mockImplementation(resolve);
                         setTimeout(resolve, 50);
                     }),
                 };
-                jest.spyOn(ProfilesUtils.extenderTypeReady, "get").mockReturnValueOnce(profilePromise);
+                jest.spyOn(ProfilesUtils.extenderTypeReady, "get").mockReturnValueOnce(profilePromise as any);
                 isProfileLockedMock.mockReturnValueOnce(false);
                 const ussApiMock = {
                     fileList: jest.fn().mockResolvedValueOnce({